Problem still persists.
Seems to be caused by the external library 'com.google.code.gson:gson:2.7@jar'. The exact same jar file exists in 'RobotCore-release'. The solution would be to remove the former; however, I'm not sure how exactly to do that, since I never added it in the first place. It's clearly part of one of the dependancies I imported to implement DL4J, so the hard part is figuring out which one it comes from, and how to exclude that part of the dependancy from the project.