Shrunk some sleep statements
Will probably save us a second. Like i said before, if we really want to speed up the auto, we'll need to change the path we take. And I dont believe it would be wise to try this until we have the latch and/or scoring mechanism on the robot.